+++ c2mon-server/c2mon-server-configuration/src/main/java/cern/c2mon/server/configuration/parser/impl/ConfigurationParserImpl.java
@@ -19,6 +19,8 @@ package cern.c2mon.server.configuration.parser.impl;
 import java.util.ArrayList;
 import java.util.List;
 
+import cern.c2mon.server.configuration.parser.exception.EntityDoesNotExistException;
+import cern.c2mon.shared.client.configuration.ConfigConstants;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -103,7 +105,18 @@ public class ConfigurationParserImpl implements ConfigurationParser {
           log.debug("Element {} (name = {}) already deleted. Detailed reason: {}", configurationEntity.getId(), configurationEntity.getName(), ex.getMessage());
         }
       } else if (configurationEntity.isUpdated()) {
-        results.add(entityFactory.updateInstance(configurationEntity));
+        try {
+          results.add(entityFactory.updateInstance(configurationEntity));
+        } catch (EntityDoesNotExistException e) {
+          ConfigurationElement missingEntity = new ConfigurationElement();
+          missingEntity.setStatus(ConfigConstants.Status.WARNING);
+          missingEntity.setEntityId(configurationEntity.getId());
+          missingEntity.setAction(ConfigConstants.Action.UPDATE);
+          missingEntity.setConfigId(-1L);
+          missingEntity.setEntity(ConfigConstants.Entity.MISSING);
+          results.add(missingEntity);
+          log.warn(e.getMessage());
+        }
       } else if (configurationEntity.isCreated()) {
         results.addAll(entityFactory.createInstance(configurationEntity));
       } else {