+++ c2mon-server/c2mon-server-elasticsearch/src/test/java/cern/c2mon/server/elasticsearch/ElasticsearchModuleIntegrationTest.java
@@ -28,6 +28,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 
 import java.io.IOException;
 
+import java.io.IOException;
+
 /**
  * @author Alban Marguet
  */