+++ CHANGELOG.md
@@ -12,6 +12,43 @@ For more details on a given release, please check also the [Milestone planning](
 ### Changed
 
 ### Fixed
+
+
+## [1.8.24] - 2017-09-05
+### Fixed
+- Server: Resolved possible deadlock in tag quality object (#166)
+- Server: Added again proper exception handling for (sub-)equipment metadata in elasticsearch data conversion (#165), which got accidentially removed by merge for issue #88
+
+
+## [1.8.23] - 2017-09-01
+### Added
+- Client API: First version of tag subscription by metadata (#88)
+- Client API: Improved Alarm configuration builder by adding more methods (#151)
+
+### Fixed
+- Server: Resolved `NullPointerException` during configuration when adding metadata with `null` value (#163)
+- Server: Configuration requests containing updates for non-existing entities will report a warning but not fail anymore
+- Documentation: Fixed dead link in the database section of the server documentation (#162)
+
+
+## 1.8.22 - BROKEN!
+Broken release!
+
+
+## 1.8.21 - BROKEN!
+Broken release!
+
+
+## 1.8.20 - BROKEN!
+Broken release!
+
+
+## 1.8.19 - BROKEN!
+Broken release!
+
+
+## [1.8.18] - 2017-08-16
+### Fixed
 - Fixed problem of alarm metadata that were wrongly stored the alarm cache object (#160). Bug got introduced with #89
 
 
@@ -139,7 +176,11 @@ This patch contains bug fixes for the DAQ layer.
 ### Removed
 - Remove of EquipmentLogger concept from DAQ Core (#56)
 
-[Unreleased]: https://gitlab.cern.ch/c2mon/c2mon/milestones/16
+
+[Unreleased]: https://gitlab.cern.ch/c2mon/c2mon/milestones/19
+[1.8.24]: https://gitlab.cern.ch/c2mon/c2mon/milestones/18
+[1.8.23]: https://gitlab.cern.ch/c2mon/c2mon/milestones/17
+[1.8.18]: https://gitlab.cern.ch/c2mon/c2mon/milestones/16
 [1.8.17]: https://gitlab.cern.ch/c2mon/c2mon/milestones/15
 [1.8.16]: https://gitlab.cern.ch/c2mon/c2mon/milestones/14
 [1.8.15]: https://gitlab.cern.ch/c2mon/c2mon/milestones/13